Extract calculatePercentage method
[MAILPOET-5091]
